Authorization added.

+++ app/Http/Controllers/TaskStatusController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 use App\TaskStatus;
 
@@ -13,54 +14,56 @@ class TaskStatusController extends Controller
         return view('task_status.index', compact('taskStatuses'));
     }
 
-    public function show(TaskStatus $taskStatus)
-    {
-        return view('task_status.show', compact('taskStatus'));
-    }
-
     public function create()
     {
+        Gate::authorize('manage-taskStatus');
+
         $taskStatus = new TaskStatus();
         return view('task_status.create', compact('taskStatus'));
     }
 
     public function store(Request $request)
     {
+        Gate::authorize('manage-taskStatus');
+
         $data = $this->validate($request, [
             'name' => 'required|unique:task_statuses',
         ]);
 
         TaskStatus::create($data);
+        flash(__('flash.task_status.store.success'))->success();
 
-        return redirect()
-            ->route('task_statuses.index')
-            ->with('success', 'Task Status created successfully');
+        return redirect()->route('task_statuses.index');
     }
 
     public function edit(TaskStatus $taskStatus)
     {
+        Gate::authorize('manage-taskStatus');
+
         return view('task_status.edit', compact('taskStatus'));
     }
 
     public function update(Request $request, TaskStatus $taskStatus)
     {
+        Gate::authorize('manage-taskStatus');
+
         $data = $this->validate($request, [
             'name' => 'required|unique:task_statuses,name',
         ]);
 
         $taskStatus->update($data);
+        flash(__('flash.task_status.update.success'));
 
-        return redirect()
-            ->route('task_statuses.index')
-            ->with('success', 'Task Status updated successfully');
+        return redirect()->route('task_statuses.index');
     }
 
     public function destroy(TaskStatus $taskStatus)
     {
-        $taskStatus->delete();
+        Gate::authorize('manage-taskStatus');
 
-        return redirect()
-            ->route('task_statuses.index')
-            ->with('success', 'Task Status removed successfully');
+        $taskStatus->delete();
+        flash(__('flash.task_status.update.destroy'));
+
+        return redirect()->route('task_statuses.index');
     }
 }
